tests: remove rest dependent tests
