+++ src/api/graphql/GetCurrentUser.query.ts
@@ -1,31 +1,5 @@
 import gql from 'graphql-tag'
 
-export interface User {
-    name: string
-    login: string
-    avatarUrl: string
-}
-
-export interface CurrentUser extends User {
-    organizations: { nodes: UserOrganization[] }
-    watching: { nodes: { name: string, owner: { avatarUrl: string, login: string } }[] }
-}
-
-export interface UserOrganization {
-    name: string
-    login: string
-    avatarUrl: string
-    membersWithRole: { nodes: User[] }
-    teams: { nodes: OrganizationTeam[] }
-}
-
-export interface OrganizationTeam {
-    name: string
-    slug: string
-    avatarUrl: string
-    members: { nodes: User[] }
-}
-
 const query = gql(`
     query GetCurrentUser {
         viewer {
