+++ tests/views/Guarded/Dashboard/RecentCommitsWithoutPullRequest.spec.ts
@@ -17,15 +17,15 @@ describe.concurrent('Recent commits without pull request', () => {
 		},
 	})
 
+	const commitStore = useCommitStore()
+	// Values that should be rendered
+	commitStore.recentBranches = [{ name: 'dummy_branch_name', owner: 'dummy_owner', repo: 'dummy_repository' }]
+
 	it('applies container identifier class', () => {
 		expect(wrapper.find(".recent-commits-without-pull-request").exists()).toBeTruthy()
 	})
 
 	it('renders the commits that do not have any pull request associated', async () => {
-		const commitStore = useCommitStore()
-		// Values that should be rendered
-		commitStore.recentBranches = [{ name: 'dummy_branch_name', owner: 'dummy_owner', repo: 'dummy_repository' }]
-
 		await wrapper.vm.$nextTick()
 		const renderedItems = wrapper.findAll('.recent-branch')
 
@@ -35,9 +35,6 @@ describe.concurrent('Recent commits without pull request', () => {
 	})
 
 	it('opens github page to create a new pull request', async () => {
-		const commitStore = useCommitStore()
-		commitStore.recentBranches = [{ name: 'dummy_branch_name', owner: 'dummy_owner', repo: 'dummy_repository' }]
-
 		await wrapper.vm.$nextTick()
 		const renderedItems = wrapper.findAll('.recent-branch')
 		const openPullRequestButton = renderedItems[0].get('.open-pull-request-button')
