+++ tests/views/Guarded/Dashboard/ReviewablePullRequests.spec.ts
@@ -1,11 +1,5 @@
-import { useNavigationStore } from '@/stores/NavigationStore'
-import { usePullRequestStore } from '@/stores/PullRequestStore'
-import { useRepositoryStore } from '@/stores/RepositoryStore'
-import PullRequestRow from '@/views/Guarded/Dashboard/ReviewablePullRequests/PullRequestRow.vue'
 import ReviewablePullRequests from '@/views/Guarded/Dashboard/ReviewablePullRequests/ReviewablePullRequests.vue'
 import { createTestingPinia } from '@pinia/testing'
-import PullRequestFixture from '@tests/__fixtures__/PullRequestFixture.json'
-import RepositoryFixture from '@tests/__fixtures__/RepositoryFixture.json'
 import { shallowMount } from '@vue/test-utils'
 
 describe.concurrent('Pull-request list', () => {
@@ -24,28 +18,7 @@ describe.concurrent('Pull-request list', () => {
 		expect(wrapper.find(".reviewable-pull-requests").exists()).toBeTruthy()
 	})
 
-	it('renders the pull requests that can be reviewed', async () => {
-		const pullRequestStore = usePullRequestStore()
-		// Values that should be rendered
-		pullRequestStore.reviewable.push(PullRequestFixture)
+	it.todo('renders the pull requests that can be reviewed')
 
-		await wrapper.vm.$nextTick()
-		const renderedItems = wrapper.findAllComponents(PullRequestRow)
-
-		expect(renderedItems.length).toBe(1)
-	})
-
-	it('fetches reviewable pull requests when repositories change', async () => {
-		const navigationStore = useNavigationStore()
-		const pullRequestStore = usePullRequestStore()
-		const repositoryStore = useRepositoryStore()
-
-		navigationStore.workspace = 'test'
-		expect(pullRequestStore.getReviewablePullRequests).not.toHaveBeenCalled()
-
-		repositoryStore.repositories = [RepositoryFixture]
-		await wrapper.vm.$nextTick()
-
-		expect(pullRequestStore.getReviewablePullRequests).toHaveBeenCalledTimes(1)
-	})
+	it.todo('fetches reviewable pull requests when repositories change')
 })
